pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.47.0",
    "nltk==3.9.1",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
